+++ src/hash_cache.rs
@@ -851,63 +851,6 @@ where
         false
     }
 
-    /// Iterates over all the entries in the [`HashCache`] allowing modifying each value.
-    ///
-    /// Key-value pairs that have existed since the invocation of the method are guaranteed to be
-    /// visited if they are not removed, however the same key-value pair can be visited more than
-    /// once if the [`HashCache`] gets resized by another thread.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use scc::HashCache;
-    ///
-    /// let hashcache: HashCache<u64, u32> = HashCache::default();
-    ///
-    /// assert!(hashcache.put(1, 0).is_ok());
-    /// assert!(hashcache.put(2, 1).is_ok());
-    ///
-    /// let mut acc = 0;
-    /// hashcache.for_each(|k, v| { acc += *k; *v = 2; });
-    /// assert_eq!(acc, 3);
-    /// assert_eq!(hashcache.read(&1, |_, v| *v).unwrap(), 2);
-    /// assert_eq!(hashcache.read(&2, |_, v| *v).unwrap(), 2);
-    /// ```
-    #[inline]
-    pub fn for_each<F: FnMut(&K, &mut V)>(&self, mut f: F) {
-        self.retain(|k, v| {
-            f(k, v);
-            true
-        });
-    }
-
-    /// Iterates over all the entries in the [`HashCache`] allowing modifying each value.
-    ///
-    /// Key-value pairs that have existed since the invocation of the method are guaranteed to be
-    /// visited if they are not removed, however the same key-value pair can be visited more than
-    /// once if the [`HashCache`] gets resized by another task.
-    ///
-    /// It is an asynchronous method returning an `impl Future` for the caller to await.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use scc::HashCache;
-    ///
-    /// let hashcache: HashCache<u64, u32> = HashCache::default();
-    ///
-    /// let future_put = hashcache.put_async(1, 0);
-    /// let future_for_each = hashcache.for_each_async(|k, v| println!("{} {}", k, v));
-    /// ```
-    #[inline]
-    pub async fn for_each_async<F: FnMut(&K, &mut V)>(&self, mut f: F) {
-        self.retain_async(|k, v| {
-            f(k, v);
-            true
-        })
-        .await;
-    }
-
     /// Retains the entries specified by the predicate.
     ///
     /// This method allows the predicate closure to modify the value field.
@@ -915,10 +858,6 @@ where
     /// Entries that have existed since the invocation of the method are guaranteed to be visited
     /// if they are not removed, however the same entry can be visited more than once if the
     /// [`HashCache`] gets resized by another thread.
-    ///
-    /// Returns `(number of remaining entries, number of removed entries)` where the number of
-    /// remaining entries can be larger than the actual number since the same entry can be visited
-    /// more than once.
     ///
     /// # Examples
     ///
@@ -931,11 +870,15 @@ where
     /// assert!(hashcache.put(2, 1).is_ok());
     /// assert!(hashcache.put(3, 2).is_ok());
     ///
-    /// assert_eq!(hashcache.retain(|k, v| *k == 1 && *v == 0), (1, 2));
+    /// hashcache.retain(|k, v| *k == 1 && *v == 0);
+    ///
+    /// assert!(hashcache.contains(&1));
+    /// assert!(!hashcache.contains(&2));
+    /// assert!(!hashcache.contains(&3));
     /// ```
     #[inline]
-    pub fn retain<F: FnMut(&K, &mut V) -> bool>(&self, mut pred: F) -> (usize, usize) {
-        self.retain_entries(|k, v| pred(k, v))
+    pub fn retain<F: FnMut(&K, &mut V) -> bool>(&self, mut pred: F) {
+        self.retain_entries(|k, v| pred(k, v));
     }
 
     /// Retains the entries specified by the predicate.
@@ -946,10 +889,7 @@ where
     /// if they are not removed, however the same entry can be visited more than once if the
     /// [`HashCache`] gets resized by another thread.
     ///
-    /// Returns `(number of remaining entries, number of removed entries)` where the number of
-    /// remaining entries can be larger than the actual number since the same entry can be visited
-    /// more than once. It is an asynchronous method returning an `impl Future` for the caller to
-    /// await.
+    /// It is an asynchronous method returning an `impl Future` for the caller to await.
     ///
     /// # Examples
     ///
@@ -962,12 +902,8 @@ where
     /// let future_retain = hashcache.retain_async(|k, v| *k == 1);
     /// ```
     #[inline]
-    pub async fn retain_async<F: FnMut(&K, &mut V) -> bool>(
-        &self,
-        mut filter: F,
-    ) -> (usize, usize) {
-        let mut num_retained: usize = 0;
-        let mut num_removed: usize = 0;
+    pub async fn retain_async<F: FnMut(&K, &mut V) -> bool>(&self, mut filter: F) {
+        let mut removed = false;
         let mut current_array_holder = self.array.get_arc(Acquire, &Barrier::new());
         while let Some(current_array) = current_array_holder.take() {
             self.cleanse_old_array_async(&current_array).await;
@@ -986,11 +922,9 @@ where
                                 let mut entry_ptr = EntryPtr::new(&barrier);
                                 while entry_ptr.next(&locker, &barrier) {
                                     let (k, v) = entry_ptr.get_mut(data_block_mut, &mut locker);
-                                    if filter(k, v) {
-                                        num_retained = num_retained.saturating_add(1);
-                                    } else {
+                                    if !filter(k, v) {
                                         locker.erase(data_block_mut, &entry_ptr);
-                                        num_removed = num_removed.saturating_add(1);
+                                        removed = true;
                                     }
                                 }
                             }
@@ -1005,18 +939,15 @@ where
                 if new_current_array.as_ptr() == current_array.as_ptr() {
                     break;
                 }
-                num_retained = 0;
                 current_array_holder.replace(new_current_array);
                 continue;
             }
             break;
         }
 
-        if num_removed >= num_retained {
+        if removed {
             self.try_resize(0, &Barrier::new());
         }
-
-        (num_retained, num_removed)
     }
 
     /// Clears all the key-value pairs.
@@ -1029,11 +960,13 @@ where
     /// let hashcache: HashCache<u64, u32> = HashCache::default();
     ///
     /// assert!(hashcache.put(1, 0).is_ok());
-    /// assert_eq!(hashcache.clear(), 1);
+    /// hashcache.clear();
+    ///
+    /// assert!(!hashcache.contains(&1));
     /// ```
     #[inline]
-    pub fn clear(&self) -> usize {
-        self.retain(|_, _| false).1
+    pub fn clear(&self) {
+        self.retain(|_, _| false);
     }
 
     /// Clears all the key-value pairs.
@@ -1051,8 +984,8 @@ where
     /// let future_clear = hashcache.clear_async();
     /// ```
     #[inline]
-    pub async fn clear_async(&self) -> usize {
-        self.retain_async(|_, _| false).await.1
+    pub async fn clear_async(&self) {
+        self.retain_async(|_, _| false).await;
     }
 
     /// Returns the number of entries in the [`HashCache`].